Misc fixes and cleanup

+++ src/main/java/org/appledash/noodel/texture/SpriteSheet.java
@@ -4,17 +4,24 @@ import java.awt.image.BufferedImage;
 
 public class SpriteSheet {
     private final Texture2D texture;
+    private final boolean textureOwned; /* if true, we made the Texture2D and we need to delete it. */
     private final int spriteWidth;
     private final int spriteHeight;
     private final int xCount;
     private final int yCount;
 
     public SpriteSheet(BufferedImage bufferedImage, int spriteWidth, int spriteHeight) {
-        this(new Texture2D(bufferedImage), spriteWidth, spriteHeight);
+        this(new Texture2D(bufferedImage), true, spriteWidth, spriteHeight);
     }
 
     public SpriteSheet(Texture2D texture, int spriteWidth, int spriteHeight) {
+        this(texture, false, spriteWidth, spriteHeight);
+    }
+
+
+    private SpriteSheet(Texture2D texture, boolean textureOwned, int spriteWidth, int spriteHeight) {
         this.texture = texture;
+        this.textureOwned = textureOwned;
         this.spriteWidth = spriteWidth;
         this.spriteHeight = spriteHeight;
         this.xCount = texture.getWidth() / spriteWidth;
@@ -42,14 +49,8 @@ public class SpriteSheet {
     }
 
     public void delete() {
-        this.texture.delete();
-    }
-
-    public static SpriteSheet fromResource(String resourcePath, int spriteWidth, int spriteHeight) {
-        return new SpriteSheet(
-                Texture2D.fromResource(resourcePath),
-                spriteWidth,
-                spriteHeight
-        );
+        if (this.textureOwned) {
+            this.texture.delete();
+        }
     }
 }
